requirements.txt
allure_python_commons==2.9.43
requests==2.25.1
pytest==6.2.4
PyYAML==5.4.1
